% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{get_dpeth_from_n}
\alias{get_dpeth_from_n}
\title{Get depth from sample size}
\usage{
get_dpeth_from_n(n, alpha = 0.05)
}
\arguments{
\item{n}{a sample size}

\item{alpha}{alpha level such as 0.1, 0.05, 0.01}
}
\value{
an integer depth
}
\description{
Calculates the depth given a sample size and alpha level
}
\examples{
get_dpeth_from_n(1e4, 0.05)
}
