% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{get_n_from_depth}
\alias{get_n_from_depth}
\title{Get sample size from depth}
\usage{
get_n_from_depth(k, alpha)
}
\arguments{
\item{k}{an integer depth}

\item{alpha}{alpha level such as 0.1, 0.05, 0.01}
}
\value{
a float sample size
}
\description{
Calculates the sample size needed given an alpha level and depth
}
\examples{
get_n_from_depth(7, 0.01)
}
