% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{make_wodd_name_big_data}
\alias{make_wodd_name_big_data}
\title{make_wodd_name_big_data}
\usage{
make_wodd_name_big_data(index)
}
\arguments{
\item{index}{int}
}
\value{
A vector
}
\description{
make_wodd_name_big_data a private function
}
