% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{raw_wodd}
\alias{raw_wodd}
\title{Raw wodd}
\usage{
raw_wodd(index)
}
\arguments{
\item{index}{- an int}
}
\value{
a vector string of names
}
\description{
private function for internal package use. Calculates raw wodd name without formatting.
}
