% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{select_wodd_name_from_table}
\alias{select_wodd_name_from_table}
\title{select_wodd_name_from_table}
\usage{
select_wodd_name_from_table(index)
}
\arguments{
\item{index}{int}
}
\value{
A vector
}
\description{
select_wodd_name_from_table a private function
}
\examples{
select_wodd_name_from_table(1L)
}
