% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{wodd_format}
\alias{wodd_format}
\title{Wodd format}
\usage{
wodd_format(wodd_name)
}
\arguments{
\item{wodd_name}{- a string}
}
\value{
a vector string of names
}
\description{
private function for internal package use. cleans up names
}
