% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{make_wodd_name_big_data}
\alias{make_wodd_name_big_data}
\title{Make wodd name big data}
\usage{
make_wodd_name_big_data(index)
}
\arguments{
\item{index}{- an int}
}
\value{
a vector string of names
}
\description{
private function for internal package use. If the data exceeds depth of 100 then calculate wodds manually
}
