% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{select_wodd_name_from_table}
\alias{select_wodd_name_from_table}
\title{Make wodd name big data}
\usage{
select_wodd_name_from_table(index)
}
\arguments{
\item{index}{- an int}
}
\value{
a vector string of names
}
\description{
private function for internal package use. This is the standard way that wodds are obtained. "Dictionary"
}
