% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{wodd_format}
\alias{wodd_format}
\title{wodd_format}
\usage{
wodd_format(wodd_name)
}
\arguments{
\item{wodd_name}{string. "S0", "S1", "M". etc}
}
\value{
A string
}
\description{
wodd_format a private function
}
