% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wodds.R
\name{wodds}
\alias{wodds}
\title{Calculate Whisker Odds aka wodds}
\usage{
wodds(
  y,
  include_tail_area = FALSE,
  include_outliers = FALSE,
  include_depth = FALSE
)
}
\arguments{
\item{y}{A vector of values}
}
\value{
A dataframe of wodds
}
\description{
Calculate Whisker Odds aka wodds
}
\examples{
wodds(rnorm(1e4, 0, 1))
}
